Phase 6: Remove old pipeline control systems
Removed three overlapping systems (~881 lines):
- LexerRegistry (src/lex/lexers/common/interface.rs)
- ParserRegistry (src/lex/parsers/common/interface.rs)
- LexPipeline (src/lex/pipeline/orchestration.rs)

All functionality now unified in PipelineExecutor with config-based
processing. No deprecation needed as this codebase is the only client.

Changes:
- Removed interface.rs files and orchestration.rs
- Recreated common/mod.rs files with minimal essential types
- Updated integration tests to remove LexPipeline dependencies
- Updated module exports in lexers.rs, parsers.rs, pipeline/mod.rs

All 559 tests passing.
