+++ src/lex/lexers.rs
@@ -42,7 +42,7 @@ pub mod linebased;
 pub mod tokens_core;
 
 pub use base_tokenization::tokenize;
-pub use common::{LexError, Lexer, LexerOutput, LexerRegistry};
+pub use common::{LexError, Lexer, LexerOutput};
 pub use detokenizer::detokenize;
 pub use tokens_core::Token;
 // Re-export line-based types for convenience
