+++ src/lex/parsers.rs
@@ -21,7 +21,7 @@ pub mod linebased;
 pub mod reference;
 
 // Re-export common parser interfaces
-pub use common::{ParseError, Parser, ParserInput, ParserRegistry};
+pub use common::{ParseError, Parser, ParserInput};
 
 // Re-export AST types and utilities from the ast module
 pub use crate::lex::ast::{
